Implement a `/healthz` endpoint for k8s readiness probe
For the time being this simply returns an empty response and a 200
status code; Additional data points may be added in the future.  It
is meant only to indicate only that the Kask endpoint is up and able
to serve requests in the most minimal way.  It does not guarantee that
requests will be successful.

Bug: T209108
Change-Id: I5cd7d0ba11114a1204d9d16523ad46b422b818e3