Serve OpenAPI specification
The approach taken here is to use an optional configuration directive,
`openapi_spec`, to specify a complete path to a YAML-formatted OpenAPI
specification.  Since Kask allows you to configure the base URI prefix
on an instance-by-instance basis, the specified file is templated with
the configured `base_uri`.  This allows each uniquely configured
instance to treat the spec as site config (if necessary; the defaults
will likely work in most cases).

The handler is fail-safe; If an error is encountered (for example if
the specified spec does not exist), the service will carry on unabated,
logging any failures. No attempt is made to validate the specification.

Bug: T222227
Change-Id: Ifcb4523aaa52d61f7ffeb487153c9c45bacf1a63