Accept a list of hostnames as Cassandra contacts
The GoCQL driver does not treat contacts as colon-delimited
hostname-and-port strings (unlike most drivers), it uses a list of
hostnames and a single port (expected to be the same for all
contacts).  That makes this change very straighforward, simply use an
array of hostnames where we formerly accepted just the one; It should
have been like this from the beginning.

Bug: T224620
Change-Id: Icf4da73c30be86949d89f5aa173cc661feb16baf