Treat Git tag as the canonical version
It took documenting a release process to see how redundant it would be
to have both a manually edited version string, and a Git version based
on the output of `git describe`; This changeset uses the output of
`describe` as the canonical version string.

With this change, the release process is reduced to:

  * Tag
  * Push

The output of `git describe` is such that after tagging a release, the
version will match the tag (say "v1.0.0").  Subsequent commits will
result in a snapshot version that is based on the most recent tag, the
number of commits on top of the tagged object, and the abbreviated
object name of the most recent commit, (for example
v1.0.0-1-gaad1c96).

Bug: T226988
Change-Id: I39039845951e5dfdc869ef09ebefbc7d12900405